GH-410: [fix] Extract shared user lookup to remove notifications require cycle
